proper error message when getting a model from a non done traintuple
When getting a model from a not done traintuple, the API returns the
followed http response (400):
`{'message': "'NoneType' object is not subscriptable"}`

This is not very informative. Add a check to ensure that the traintuple
is at a correct status.

On top of that add exception logs for easier debugging (all types of exception
are catched and returned to the API user as a string without any
log/print/traceback in the server).