Add base image cache
Adding cache to kaniko to avoid dowloading the same base image every time we biuild a container seems to speed up the build time.
More tests are passing and not throwing a timeout.
This is still slow compared to a docker build as we can't use cache from every layer and just cache the base image.