[doc] expose generated openapi spec
This adds 3 new routes:
- /schema/ exposing the openapi spec
- /schema/swagger-ui/ to browse the spec with the swagger UI
- /schema/redoc/ to browse the spec with the redoc UI

You have to be authenticated to acces those routes.

A similar feature previously existed, provided by django-rest-swagger
but the repository has been archived and the author recommended
drf-yasg. On the documentation page of drf-yasg, the author mentions
that drf-yasg won't support openapi 3.0 and recommend to use
drf-spectacular to expose an openapi 3.0 spec.

That's why I chose drf-spectacular, which works the same way than
django-rest-framework and is BSD licensed.