Fix the celery event key for compute plan events
Historically, the ledger was configured to only have one transaction per block. However, we changed this pradigm a few months back when we changed the ledger configuration to allow multiple transactions per block. As a result, the block_id is not enough to identify a transaction. As a result, when a single block contains multiple transactions with events, only the first event is processed, and the other ones are skipped. This commit changes the key to uniquely identify each event/transaction.