refactor: organization client (#1253)
I merged the two `http_post` and `http_get` functions in a single one. I also extracted the settings and headers and kwargs building code. Some change was also done to view utils that was previously using a low level function to have a cleaner contract between the API code and the client.