fix(chart): add imagePullSecrets to migrations job (#1271)
This addresses an issue we observed with Florian Jouan when working on NestCC, where we were met with an image pull failure due to a lack of permissions. Adding the image pull secrets to the job via editing the job seemed to fix the issue.

My hypothesis is this isn't a problem usually, because we generally deploy Connect on single-node k8s clusters. By default K8s will just use the specified image if it is already present on the node, so as long as the server has the creds the migration job could use the image. But NestCC has two nodes, and we were unlucky with the scheduling.