chore: use mypy to type check part of the compute engine code (#1277)
Apply Mypy on the `/backend/substrapp/tasks` code with a custom config. The goal of this config is to ensure all functions within this module are always type hinted. Currently we don't type check dependencies as some of our code is still not type hinted enough and that would raise too many errors.

The Mypy check was added to the lint make entry and the CI.