refactor: orchestrator client return algo objects (#1272)
In this PR I removed the `Algo` class I introduced at some point and replaced it with `orchestrator.Algo`. It makes mocking easier. The only drawback was that I had to do the query to retrieve the algo asset in the image builder. That's the reason I introduced the `Datastore` class. Doing this it is easier to mock the asset retrieval and open the possibility for another kind of storage without having to change all our code.