feat: delete transient outputs (#1279)
We want to delete assets from the backend when they are not used anymore. Previously this was done using a Celery task that deleted models from the database before removing them from the orchestrator. Since now with Generic tasks outputs are not limited to models anymore we need to handle this deletion in a more generic way.

Add two new celery tasks:
- `remove_transient_outputs_from_orc` to call the `DisableOutputs` RPC from the orchestrator on all outputs that can be disabled.
- `remove_intermediary_models_from_db_new` to delete models from the DB only. The `new` will be removed later but for now we need to preserve compatibility with the previous compute plan flag for model deletion. Note that this will probably raise some errors in the new task because we will try to delete models that are already deleted.

Add a new file `backend/events/handler_compute_engine.py` I did this to keep the reactor readable and as free as possible from compute engine related code.

This PR only add code but we should see the benefit soon when the code related to the previous compute plan flag handling will be removed.